+++ src/broadcast.cu
@@ -31,7 +31,7 @@ testResult_t BroadcastInitData(struct threadArgs* args, ncclDataType_t type, ncc
   return testSuccess;
 }
 
-void BroadcastGetBw(size_t count, int typesize, double sec, double* algBw, double* busBw, int nranks) {
+void BroadcastGetBw(size_t count, size_t typesize, double sec, double* algBw, double* busBw, int nranks) {
   double baseBw = (double)(count * typesize) / 1.0E9 / sec;
 
   *algBw = baseBw;
